Bugfix: Z_Piece rotation and OutOfBounds error.
Z_Piece was not rotating correctly in the 270 degree position, offsets
were not correctly set. I fixed the offsets.

When rotating a piece, if the rotation caused the piece to fall out of
the board, or on top of another piece, it made the whole game crash. I
fixed it by checking if the final position was legal, and preventing
the rotation otherwise.